fix: 점수 타입 Double로 수정, 모의고사 문제 구조에 맞게 필드 변경

+++ src/main/java/web/tosunsaeng/domain/exams/domain/entity/ExamResult.java
@@ -33,8 +33,8 @@ public class ExamResult {
     // 문항 데이터 (요약 데이터일 때는 null)
     private Integer partNumber;
     private Integer questionNumber;
-    private Integer score;
-    private Integer maxScore;
+    private Double score;
+    private Double maxScore;
     private String transcript;
     private ItemFeedback feedback;
